Split config getting
