+++ lib/has-status.js
@@ -1,0 +1,7 @@
+"use strict";
+Object.defineProperty(exports, "__esModule", { value: true });
+exports.hasStatus = void 0;
+function hasStatus(obj) {
+    return Object.prototype.hasOwnProperty.call(obj, "status");
+}
+exports.hasStatus = hasStatus;
