+++ lib/tested-version.js
@@ -10,54 +10,22 @@ var __awaiter = (this && this.__awaiter) || function (thisArg, _arguments, P, ge
 };
 Object.defineProperty(exports, "__esModule", { value: true });
 exports.testedVersion = void 0;
-const Config_1 = require("./interfaces/Config");
 const octokit_1 = require("./octokit");
 const repo_1 = require("./repo");
+const wpvc_config_1 = require("./wpvc-config");
+const has_status_1 = require("./has-status");
 const ConfigError_1 = require("./exceptions/ConfigError");
 const InvalidReadmeError_1 = require("./exceptions/InvalidReadmeError");
-function hasStatus(obj) {
-    return Object.prototype.hasOwnProperty.call(obj, "status");
-}
-function WPVCConfig() {
-    return __awaiter(this, void 0, void 0, function* () {
-        const file = yield octokit_1.octokit.repos.getContent(Object.assign(Object.assign({}, repo_1.repo), { path: '.wordpress-version-checker.json' })).catch(function (e) {
-            if (hasStatus(e) && e.status === 404) {
-                return null;
-            }
-            else {
-                throw new ConfigError_1.ConfigError(String(e));
-            }
-        });
-        if (file === null) {
-            return null;
-        }
-        const encodedContent = file.data.content;
-        if (!encodedContent) {
-            throw new ConfigError_1.ConfigError('Failed to decode the file.');
-        }
-        let config = {};
-        try {
-            config = JSON.parse(Buffer.from(encodedContent, 'base64').toString());
-        }
-        catch (e) {
-            throw new ConfigError_1.ConfigError(e.message);
-        }
-        if (!Config_1.isConfig(config)) {
-            throw new ConfigError_1.ConfigError('Invalid config file.');
-        }
-        return config;
-    });
-}
 function readme() {
     return __awaiter(this, void 0, void 0, function* () {
         let readmeLocations = ['readme.txt', 'plugin/readme.txt'];
-        const config = yield WPVCConfig();
+        const config = yield wpvc_config_1.WPVCConfig();
         if (config !== null) {
             readmeLocations = [config.readme];
         }
         for (const readmeLocation of readmeLocations) {
             const result = yield octokit_1.octokit.repos.getContent(Object.assign(Object.assign({}, repo_1.repo), { path: readmeLocation })).catch(function (e) {
-                if (hasStatus(e) && e.status === 404) {
+                if (has_status_1.hasStatus(e) && e.status === 404) {
                     return null;
                 }
                 else {
