+++ src/has-status.ts
@@ -1,0 +1,3 @@
+export function hasStatus(obj: Record<string, unknown>): obj is Record<"status", unknown> {
+	return Object.prototype.hasOwnProperty.call(obj, "status")
+}
